clean up program metadata idl fetching hooks
