+++ app/address/[address]/idl/page.tsx
@@ -1,16 +1,24 @@
 'use client';
 
 import { IdlCard } from '@components/account/IdlCard';
-import { useAnchorProgram } from '@providers/anchor';
+import { useIdlFromAnchorProgramSeed } from '@providers/anchor';
 import { useCluster } from '@providers/cluster';
-import { useIdlFromProgramMetadataProgram } from '@providers/idl';
-import { useState } from 'react';
+import { useIdlFromMetadataProgram } from '@providers/idl';
+import { Suspense, useEffect, useState } from 'react';
 
 export default function IdlPage({ params: { address } }: { params: { address: string } }) {
     const { url } = useCluster();
+    const anchorIdl = useIdlFromAnchorProgramSeed(address, url, false);
+    const metadataIdl = useIdlFromMetadataProgram(address, url, false);
+
     const [activeTab, setActiveTab] = useState<'anchor' | 'metadata'>('anchor');
-    const { idl: anchorIdl } = useAnchorProgram(address, url);
-    const { idl: metadataIdl } = useIdlFromProgramMetadataProgram(address, url);
+
+    useEffect(() => {
+        // Show whatever tab is available
+        if (!anchorIdl && metadataIdl) {
+            setActiveTab('metadata');
+        }
+    }, [anchorIdl, metadataIdl]);
 
     return (
         <div className="card">
@@ -39,13 +47,25 @@ export default function IdlPage({ params: { address } }: { params: { address: st
                 </ul>
             </div>
             <div className="card-body">
-                {activeTab === 'anchor' && anchorIdl && (
-                    <IdlCard idl={anchorIdl} programId={address} title="Anchor IDL" />
-                )}
-                {activeTab === 'metadata' && metadataIdl && (
-                    <IdlCard idl={metadataIdl} programId={address} title="Anchor IDL (Program metadata)" />
-                )}
+                <Suspense fallback={<div>Loading...</div>}>
+                    {activeTab === 'anchor' && anchorIdl && <AnchorIdlCard programId={address} url={url} />}
+                    {activeTab === 'metadata' && metadataIdl && (
+                        <ProgramMetadataIdlCard url={url} programId={address} />
+                    )}
+                </Suspense>
             </div>
         </div>
     );
+}
+
+function ProgramMetadataIdlCard({ programId, url }: { programId: string; url: string }) {
+    const idl = useIdlFromMetadataProgram(programId, url, true);
+
+    return <IdlCard idl={idl ?? ({} as any)} programId={programId} title="Program Metadata IDL" />;
+}
+
+function AnchorIdlCard({ programId, url }: { programId: string; url: string }) {
+    const idl = useIdlFromAnchorProgramSeed(programId, url, true);
+
+    return <IdlCard idl={idl ?? ({} as any)} programId={programId} title="Anchor IDL" />;
 }
